--- app/controllers/conversions/tasks_controller.rb
@@ -1,6 +1,0 @@
-class Conversions::TasksController < ApplicationController
-  def index
-    @project = Project.find(params[:project_id])
-    @task_list = Conversion::TaskList.new(@project, current_user)
-  end
-end
