+++ app/controllers/project_information_controller.rb
@@ -1,4 +1,6 @@
 class ProjectInformationController < ApplicationController
+  include Projectable
+
   def show
     @project = Project.find(params[:project_id])
   end
