+++ app/controllers/projects_controller.rb
@@ -6,13 +6,7 @@ class ProjectsController < ApplicationController
     @project = Project.find(params[:id])
     authorize @project
 
-    # TODO: Temporary until we add Tasks for Transfer Projects
-    case @project.type
-    when "Conversion::Project"
-      redirect_to project_conversion_tasks_path(@project)
-    when "Transfer::Project"
-      render "transfers/projects/show"
-    end
+    redirect_to project_tasks_path(@project)
   end
 
   def index
