+++ app/views/conversions/tasks/check_accuracy_of_higher_needs/edit.html.erb
@@ -1,7 +1,7 @@
 <%= render(TaskList::TaskHeaderComponent.new(project: @project, task: @task)) %>
 
 <% content_for :pre_content_nav do %>
-  <% render partial: "shared/back_link", locals: {href: project_conversion_tasks_path(@project)} %>
+  <% render partial: "shared/back_link", locals: {href: project_tasks_path(@project)} %>
 <% end %>
 
 <div class="govuk-grid-row">
