+++ app/views/conversions/tasks/shared/_back_link.erb
@@ -1,3 +1,3 @@
 <nav aria-label="back link">
-  <%= govuk_back_link(href: project_conversion_tasks_path(project_id)) %>
+  <%= govuk_back_link(href: project_tasks_path(project_id)) %>
 </nav>
