+++ app/views/projects/shared/_project_sub_navigation.html.erb
@@ -1,7 +1,7 @@
 <nav class="moj-sub-navigation" aria-label="Project sub-navigation">
   <ul class="moj-sub-navigation__list">
 
-    <%= render partial: "shared/sub_navigation_item", locals: {name: t("subnavigation.project_task_list"), path: project_conversion_tasks_path(@project)} %>
+    <%= render partial: "shared/sub_navigation_item", locals: {name: t("subnavigation.project_task_list"), path: project_tasks_path(@project)} %>
 
     <%= render partial: "shared/sub_navigation_item", locals: {name: t("subnavigation.project_information"), path: project_information_path(@project)} %>
