+++ spec/features/conversions/users_can_select_a_funding_agreement_letter_contact_spec.rb
@@ -14,7 +14,7 @@ RSpec.feature "Users select a contact to receive the funding agreement letters" 
     let!(:contact_2) { create(:project_contact, project: project, name: "Jane Jones") }
 
     it "allows the user to select one of the existing contacts on the Funding Agreement Letters task page" do
-      visit project_conversion_tasks_path(project)
+      visit project_tasks_path(project)
       click_on "Confirm who will get the funding agreement letters"
       expect(page).to have_content(contact_1.name)
       expect(page).to have_content(contact_2.name)
@@ -36,7 +36,7 @@ RSpec.feature "Users select a contact to receive the funding agreement letters" 
     end
 
     it "shows the contact as preselected on the Funding Agreement Letters task page" do
-      visit project_conversion_tasks_path(project)
+      visit project_tasks_path(project)
       click_on "Confirm who will get the funding agreement letters"
 
       expect(page).to have_checked_field(contact_2.name)
@@ -44,7 +44,7 @@ RSpec.feature "Users select a contact to receive the funding agreement letters" 
 
     context "and the funding agreement letters contact is changed" do
       it "switches the contact for the project" do
-        visit project_conversion_tasks_path(project)
+        visit project_tasks_path(project)
         click_on "Confirm who will get the funding agreement letters"
 
         choose contact_1.name
@@ -57,7 +57,7 @@ RSpec.feature "Users select a contact to receive the funding agreement letters" 
 
   context "when the project does not have any contacts" do
     it "directs the user to the external contacts page" do
-      visit project_conversion_tasks_path(project)
+      visit project_tasks_path(project)
       click_on "Confirm who will get the funding agreement letters"
 
       expect(page).to have_content("Add contacts")
@@ -66,7 +66,7 @@ RSpec.feature "Users select a contact to receive the funding agreement letters" 
     end
 
     it "allows the user to go back to the task list without adding a user" do
-      visit project_conversion_tasks_path(project)
+      visit project_tasks_path(project)
       click_on "Confirm who will get the funding agreement letters"
 
       expect(page).to have_content("Add contacts")
